requirements.txt
streamlit==1.40.2
groq==0.13.0
python-dotenv==1.0.1
